fix solidity Format

+++ apps/contracts/contracts/Zengo.sol
@@ -11,6 +11,7 @@ import "./storage/ZengoStorage.sol";
 
 contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, ContractMetadata {
     IERC20 public token;
+
     event ModeratorsAdded(address[] indexed newModerators);
     event ModeratorRemoved(address indexed removedModerator);
     // nested mapping cannot be emitted in events
@@ -22,25 +23,17 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
     }
 
     modifier onlyModerator() {
-        require(
-            moderators[msg.sender],
-            "Only moderators can call this function"
-        );
+        require(moderators[msg.sender], "Only moderators can call this function");
         _;
     }
 
     modifier onlyProposer(uint256 _proposalId) {
-        require(proposals[_proposalId].proposer == msg.sender,
-            "Only proposer of this proposal can call this function"
-        );
+        require(proposals[_proposalId].proposer == msg.sender, "Only proposer of this proposal can call this function");
         _;
     }
 
-    constructor(
-        uint256 _pluralVotingPoints,
-        address _tokenAddress
-    ) {
-        owner = msg.sender;   
+    constructor(uint256 _pluralVotingPoints, address _tokenAddress) {
+        owner = msg.sender;
 
         pluralVotingPoints = _pluralVotingPoints;
         token = IERC20(_tokenAddress);
@@ -48,13 +41,23 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
         moderatorList.push(msg.sender);
     }
 
-    function updateModeratorInfo(address _moderatorAddress, uint8 _moderatorType, string memory _position, string memory _organization) external onlyOwner {
+    function updateModeratorInfo(
+        address _moderatorAddress,
+        uint8 _moderatorType,
+        string memory _position,
+        string memory _organization
+    ) external onlyOwner {
         moderatorStruct[_moderatorAddress].moderatorType = Structs.ModeratorType(_moderatorType);
         moderatorStruct[_moderatorAddress].position = _position;
         moderatorStruct[_moderatorAddress].organization = _organization;
     }
-    
-    function addModerator(address _moderatorAddress, uint8 _moderatorType, string memory _position, string memory _organization) external onlyOwner checkState(0){
+
+    function addModerator(
+        address _moderatorAddress,
+        uint8 _moderatorType,
+        string memory _position,
+        string memory _organization
+    ) external onlyOwner checkState(0) {
         require(moderators[_moderatorAddress] == false, "Address is already a Moderator");
         moderators[_moderatorAddress] = true;
         moderatorList.push(_moderatorAddress);
@@ -64,7 +67,7 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
     }
 
     function addModerators(address[] memory _moderators) external onlyOwner checkState(0) {
-        for (uint i = 0; i < _moderators.length; i++) {
+        for (uint256 i = 0; i < _moderators.length; i++) {
             if (moderators[_moderators[i]] == false) {
                 moderatorList.push(_moderators[i]);
                 moderators[_moderators[i]] = true;
@@ -85,7 +88,6 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
         emit ModeratorRemoved(_moderator);
     }
 
-
     // TODO: constructing a constructor like this
     // doesn't work figure out how to write it
     // Done
@@ -98,7 +100,7 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
         string memory _streetAddress,
         uint256 _latitude,
         uint256 _longitude
-    ) external checkState(1){
+    ) external checkState(1) {
         // require(
         //     votingPoints[msg.sender] > 0,
         //     "You need voting points to submit a proposal"
@@ -109,13 +111,13 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
         // });
 
         Structs.Proposal storage newProposal = proposals[proposalCount];
-        // Structs.Proposal storage newProposal = 
+        // Structs.Proposal storage newProposal =
 
-        Structs.Evidence memory newEvidence = Structs.Evidence(_evidenceDescription, _streetAddress, _evidenceUri, _latitude, _longitude);
+        Structs.Evidence memory newEvidence =
+            Structs.Evidence(_evidenceDescription, _streetAddress, _evidenceUri, _latitude, _longitude);
 
         // proposals[proposalCount] = Structs.Proposal(0, proposalCount, _title, _proposalDescription, _proposalType, msg.sender, newEvidence, [], Structs.VerificationState(0), false, false);
 
-        
         newProposal.proposalId = proposalCount;
         newProposal.title = _title;
         newProposal.proposalType = _proposalType;
@@ -135,11 +137,7 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
         proposalCount++;
     }
 
-
-    function setIndividualVotingPoints(
-        address _voter,
-        uint256 _points
-    ) external onlyOwner {
+    function setIndividualVotingPoints(address _voter, uint256 _points) external onlyOwner {
         votingPoints[_voter] = _points;
     }
 
@@ -149,25 +147,16 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
         }
     }
 
-    function _canSetContractURI()
-        internal
-        view
-        virtual
-        override
-        returns (bool)
-    {
+    function _canSetContractURI() internal view virtual override returns (bool) {
         return msg.sender == deployer;
     }
 
-    function voteToClassifyProposal(
-        uint8 _vote,
-        uint8 _votingIteration,
-        uint256 _proposalId
-    ) external onlyModerator checkState(1){
-        require(
-            _vote <= uint8(Structs.VerificationState.ApproveForFunding),
-            "Out of Range / Invalid vote option"
-        );
+    function voteToClassifyProposal(uint8 _vote, uint8 _votingIteration, uint256 _proposalId)
+        external
+        onlyModerator
+        checkState(1)
+    {
+        require(_vote <= uint8(Structs.VerificationState.ApproveForFunding), "Out of Range / Invalid vote option");
 
         require(
             !hasVoted[_proposalId][_votingIteration][msg.sender],
@@ -183,23 +172,20 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
         // consensusIteration reaches the threshold votesPercent
         // update addModerator flag here
         if (
-            (voteCount[_proposalId][_votingIteration][Structs.VerificationState(_vote)] * 100) >
-            moderatorList.length * THRESHOLD_VOTE_LIMIT
+            (voteCount[_proposalId][_votingIteration][Structs.VerificationState(_vote)] * 100)
+                > moderatorList.length * THRESHOLD_VOTE_LIMIT
         ) {
-            autoTriggerVoteResult(
-                _votingIteration,
-                _proposalId,
-                Structs.VerificationState(_vote)
-            );
+            autoTriggerVoteResult(_votingIteration, _proposalId, Structs.VerificationState(_vote));
         }
     }
 
-    function autoTriggerVoteResult(
-        uint8 _votingIteration,
-        uint256 _proposalId,
-        Structs.VerificationState _resultState
-    ) internal {
-        if (_resultState == Structs.VerificationState(1) || _resultState == Structs.VerificationState(2) || _resultState == Structs.VerificationState(3)) {
+    function autoTriggerVoteResult(uint8 _votingIteration, uint256 _proposalId, Structs.VerificationState _resultState)
+        internal
+    {
+        if (
+            _resultState == Structs.VerificationState(1) || _resultState == Structs.VerificationState(2)
+                || _resultState == Structs.VerificationState(3)
+        ) {
             voteIterations[_proposalId][_votingIteration].inProgress = true;
             voteIterations[_proposalId][_votingIteration].resultState = _resultState;
             // TODO: emit Event that proposal is now respective
@@ -221,7 +207,7 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
             proposals[_proposalId].isVerified = false;
             // TODO: emit Event that proposal is rejected or spam and
             // is ineligible for funding
-        } else if (_resultState ==Structs.VerificationState(6)) {
+        } else if (_resultState == Structs.VerificationState(6)) {
             voteIterations[_proposalId][_votingIteration].inProgress = false;
             voteIterations[_proposalId][_votingIteration].resultState = _resultState;
             proposals[_proposalId].isEligibleForFunding = true;
@@ -231,10 +217,11 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
     }
 
     // Should this be callable by moderator?
-    function concludeVotingIteration(
-        uint8 _votingIteration,
-        uint256 _proposalId
-    ) external onlyModerator checkState(1){
+    function concludeVotingIteration(uint8 _votingIteration, uint256 _proposalId)
+        external
+        onlyModerator
+        checkState(1)
+    {
         // TODO: check global State
         // TODO: check zero Votes
         require(
@@ -246,9 +233,7 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
         uint256 maxCount = 0;
         // Have to handle Edge case when there are equal number of Votes
         for (uint8 i = 0; i < 6; i++) {
-            if (
-                voteCount[_proposalId][_votingIteration][Structs.VerificationState(i)] > maxCount
-            ) {
+            if (voteCount[_proposalId][_votingIteration][Structs.VerificationState(i)] > maxCount) {
                 result = i;
             }
         }
@@ -284,7 +269,6 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
     }
 
     function intializeVotingIteration(uint256 _proposalId) internal {
-
         // uint256 memory length = proposals[_proposalId].votingIterations.length;
         voteIterations[_proposalId].push();
 
@@ -299,8 +283,7 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
     }
 
     function addVotingIteration(uint256 _proposalId) public onlyModerator checkState(1) {
-        uint8 currentVoteIteration = proposals[_proposalId]
-            .votingIterationCount;
+        uint8 currentVoteIteration = proposals[_proposalId].votingIterationCount;
 
         // uint256 memory length = proposals[_proposalId].votingIterations.length;
         voteIterations[_proposalId].push();
@@ -316,12 +299,20 @@ contract ZengoDAO is Constants, ZengoStorage, GStates, PermissionsEnumerable, Co
     }
 
     //  TODO: Add function to add Evidence to Voting Iteration
-    function addEvidence(uint256 _proposalId, uint8 _votingIteration, uint256 _latitude, uint256 _longitude, string memory _evidenceDescription, string memory _streetAddress, string memory _evidenceUri) public onlyProposer(_proposalId) {
-        
+    function addEvidence(
+        uint256 _proposalId,
+        uint8 _votingIteration,
+        uint256 _latitude,
+        uint256 _longitude,
+        string memory _evidenceDescription,
+        string memory _streetAddress,
+        string memory _evidenceUri
+    ) public onlyProposer(_proposalId) {
         uint256 currentEvidenceIndex = votingIterationEvidence[_proposalId][_votingIteration].length;
         voteIterations[_proposalId][_votingIteration].evidenceCount++;
         votingIterationEvidence[_proposalId][_votingIteration].push();
-        votingIterationEvidence[_proposalId][_votingIteration][currentEvidenceIndex].evidenceDescription = _evidenceDescription;
+        votingIterationEvidence[_proposalId][_votingIteration][currentEvidenceIndex].evidenceDescription =
+            _evidenceDescription;
         votingIterationEvidence[_proposalId][_votingIteration][currentEvidenceIndex].streetAddress = _streetAddress;
         votingIterationEvidence[_proposalId][_votingIteration][currentEvidenceIndex].evidenceUri = _evidenceUri;
         votingIterationEvidence[_proposalId][_votingIteration][currentEvidenceIndex].latitude = _latitude;
